pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "hr-capacity-model"
version = "0.1.0"
description = "Decision-grade HR capacity pressure model with governance artifacts and reproducible runs."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Ryan Brush" }]
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-PyYAML>=6.0.12.20240808",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hr_capacity_model"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["hr_capacity_model"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
mypy_path = "src"
